refactor: extract AppNameProvider to its own module BED-8743
